+++ lib/aruba/processes/null_process.rb
@@ -3,6 +3,10 @@ require 'aruba/processes/basic_process'
 module Aruba
   module Processes
     # Null Process
+    #
+    # `NullProcess` is not meant for direct use - `BasicProcess.new` - by users.
+    #
+    # @private
     class NullProcess < BasicProcess
       def self.match?(mode)
         mode == :null